refactor(app): extract buffer-close and tab-navigation cluster (phase 5b)
Move close_buffer, force_close_buffer, close_buffer_internal (the big
234-line method), the six close_*_tab* variants, set_batch_close_status_message,
close_tab_in_split_silent, switch_buffer, next_buffer, prev_buffer,
cycle_tab, navigate_back, navigate_forward out of buffer_management.rs
into a new app/buffer_close.rs sibling. ~775 lines.

Closing a buffer in this editor coordinates: removing the buffer from
the registry, cleaning up LSP state and semantic tokens, picking the
next buffer to focus via the focus-history LRU, adjusting split tab
lists, and (for terminal buffers) tearing down the terminal manager.
Tab navigation depends on the same focus-history machinery so it
moves with it.

buffer_management.rs goes from 2972 -> 2196 lines (-776).
2026 tests still pass.
